external modules: improved predictive module
